requirements.txt
streamlit
opencv-python
pytesseract
pyzbar
requests
Pillow